we now have a deterministic way of getting to a good approximation
Specifically, we use the strongly regularised triangle as a first guess and, selecting for the largest contributions to the pmfs, we compute the spectrum on the weakly regularised triangle with a constrained number of modes. Knowing the target via DFFT, we can iterate, i.e. increase or decrease the number of modes, until the error is within approx. 10 percent of the true value. However, the issue now is this presupposes that we know the truth. If not, our first guesses are still way off, up to 100 percent difference from the true value.