orog_delaunay: improved results
using rect(FF)+tri(FF), we now get an average of -27% in the error on the approx R2B4 (200km x 130km) grid, -25% error on approximately square (120km x 120km) grid, and -10% on the finer scale grid (85km x 60km). A 5km Gaussian filter is applied to the orography in all cases.