made two major changes to utils.py; see the extended commit message
First change: The initalisation of the polygon (triangle) object uses a rescaling that brings everything between [0,1]. However, this no longer makes sense once we implement the tapering, as the triangle is now embedded in padded rectangular domain, and so the vertices of the triangles, when rescaled to [0,1], are no longer at the extrema of the boundaries. Instead, the rescaling has to consider the full extent of the surrounding rectangular domain and scale the vertices correspondingly. This has now been implemented in the changes in gen_triangle, rescale, and get_lat_lon_segment. Specifically, the initialisation of the triangle object has now been moved into the get_lat_lon_segment function. Second change: The stencil in the tapering class now uses a generic 9-point formulation.